+++ spec/rubocop/cop/rspec/unused_let/scope_spec.rb
@@ -12,22 +12,44 @@ RSpec.describe RuboCop::Cop::RSpec::UnusedLet::Scope do
 
     let(:scope) { build_scope }
 
-    before { scope.add_definition(:let, :value, def_node) }
+    context "with a definition of one name" do
+      before { scope.add_definition(:let, :value, def_node) }
 
-    context "when the definition is never marked referenced" do
-      it { is_expected.to contain_exactly([:let, :value, def_node]) }
+      context "when it is never marked referenced" do
+        it { is_expected.to contain_exactly(have_attributes(helper: :let, name: :value, node: def_node)) }
+      end
+
+      context "when the name has been marked referenced" do
+        before { scope.mark_referenced(:value) }
+
+        it { is_expected.to be_empty }
+      end
+
+      context "when only a different name has been marked referenced" do
+        before { scope.mark_referenced(:other) }
+
+        it { is_expected.to contain_exactly(have_attributes(helper: :let, name: :value, node: def_node)) }
+      end
     end
 
-    context "when the name has been marked referenced" do
-      before { scope.mark_referenced(:value) }
+    context "with an aliased definition" do
+      before { scope.add_definition(:subject, :widget, def_node, alias_name: :subject) }
 
-      it { is_expected.to be_empty }
-    end
+      context "when only the alias has been marked referenced" do
+        before { scope.mark_referenced(:subject) }
 
-    context "when only a different name has been marked referenced" do
-      before { scope.mark_referenced(:other) }
+        it { is_expected.to be_empty }
+      end
 
-      it { is_expected.to contain_exactly([:let, :value, def_node]) }
+      context "when only the declared name has been marked referenced" do
+        before { scope.mark_referenced(:widget) }
+
+        it { is_expected.to be_empty }
+      end
+
+      context "when neither name has been marked referenced" do
+        it { is_expected.to contain_exactly(have_attributes(helper: :subject, name: :widget)) }
+      end
     end
   end
 
@@ -39,9 +61,10 @@ RSpec.describe RuboCop::Cop::RSpec::UnusedLet::Scope do
     before do
       scope.add_definition(:let, :first, :first_node)
       scope.add_definition(:let!, :second, :second_node)
+      scope.add_definition(:subject, :third, :third_node, alias_name: :subject)
     end
 
-    it { is_expected.to eq(%i[first second]) }
+    it { is_expected.to eq(%i[first second third subject]) }
   end
 
   describe "#shared?" do